# backend/src/app/api/routes/users.py
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Depends, Path, Query, Response, status

from app.dependencies.auth import CurrentUser, require_superuser
from app.dependencies.user import UserServiceDep
from app.schemas.common import ApiResponse, ErrorResponse, PageData, PaginationQuery
from app.schemas.user import (
    UserCreateRequest,
    UserData,
    UserPasswordChangeRequest,
    UserPutRequest,
    UserSelfPutRequest,
)

router = APIRouter(
    prefix="/users",
    tags=["users"],
    responses={
        400: {"model": ErrorResponse},
        403: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        404: {"model": ErrorResponse},
        409: {"model": ErrorResponse},
        422: {"model": ErrorResponse},
    },
)
admin_router = APIRouter(
    prefix="/users",
    tags=["users"],
    dependencies=[Depends(require_superuser)],
    responses={
        401: {"model": ErrorResponse},
        403: {"model": ErrorResponse},
        404: {"model": ErrorResponse},
        409: {"model": ErrorResponse},
        422: {"model": ErrorResponse},
    },
)
# 路由与 OpenAPI 使用 camelCase, 函数内部仍保留 Python 的 snake_case 命名。
UserIdPath = Annotated[UUID, Path(alias="userId")]


@router.get("/me")
def get_current_user(current_user: CurrentUser) -> ApiResponse[UserData]:
    return ApiResponse(data=current_user)


@router.put("/me")
def update_current_user(
    request: UserSelfPutRequest,
    current_user: CurrentUser,
    service: UserServiceDep,
) -> ApiResponse[UserData]:
    user = service.update_current_user(
        actor=current_user,
        email=str(request.email),
        full_name=request.full_name,
    )
    return ApiResponse(data=user)


@router.put("/me/password", status_code=status.HTTP_204_NO_CONTENT)
def change_current_user_password(
    request: UserPasswordChangeRequest,
    current_user: CurrentUser,
    service: UserServiceDep,
) -> Response:
    service.change_current_user_password(
        actor=current_user,
        current_password=request.current_password,
        new_password=request.new_password,
    )
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@admin_router.post("", status_code=status.HTTP_201_CREATED)
def create_user(
    request: UserCreateRequest,
    service: UserServiceDep,
) -> ApiResponse[UserData]:
    user = service.create_user(
        email=str(request.email),
        full_name=request.full_name,
        password=request.password,
        is_active=request.is_active,
        is_superuser=request.is_superuser,
    )
    return ApiResponse(data=user)


@admin_router.get("")
def list_users(
    query: Annotated[PaginationQuery, Query()],
    service: UserServiceDep,
) -> ApiResponse[PageData[UserData]]:
    page = service.list_users(
        page=query.page,
        page_size=query.page_size,
    )
    return ApiResponse(data=page)


@admin_router.get("/{userId}")
def get_user(
    user_id: UserIdPath,
    service: UserServiceDep,
) -> ApiResponse[UserData]:
    user = service.get_user(user_id=user_id)
    return ApiResponse(data=user)


@admin_router.put("/{userId}")
def update_user(
    user_id: UserIdPath,
    request: UserPutRequest,
    current_user: CurrentUser,
    service: UserServiceDep,
) -> ApiResponse[UserData]:
    user = service.update_user_as_admin(
        actor_id=current_user.id,
        user_id=user_id,
        email=str(request.email),
        full_name=request.full_name,
        password=request.password,
        is_active=request.is_active,
        is_superuser=request.is_superuser,
    )
    return ApiResponse(data=user)


@admin_router.delete("/{userId}", status_code=status.HTTP_204_NO_CONTENT)
def delete_user(
    user_id: UserIdPath,
    current_user: CurrentUser,
    service: UserServiceDep,
) -> Response:
    service.delete_user(
        actor_id=current_user.id,
        user_id=user_id,
    )
    return Response(status_code=status.HTTP_204_NO_CONTENT)